factored out immich class
