+++ .gitignore
@@ -1,2 +1,3 @@
 downloads
 .venv
+__pycache__
